+++ src/BlockHorizons/BlockSniper/ui/windows/BrushPropertiesWindow.php
@@ -24,7 +24,7 @@ class BrushPropertiesWindow extends CustomWindow{
 				$b->size = (int) $value;
 			});
 		}
-		if($b->mode === Brush::MODE_BRUSH && $b->getType()->usesSize()){
+		if($b->mode === Brush::MODE_BRUSH && $b->getType()->usesSize() && !$b->getShape()->usesThreeLengths()){
 			$this->addToggle($this->t(Translation::UI_BRUSH_MENU_DECREMENT), $b->decrementing, function(Player $player, bool $value) use ($b){
 				$b->decrementing = $value;
 				// Set the size the brush will reset to after reaching a size of 0.
@@ -36,9 +36,17 @@ class BrushPropertiesWindow extends CustomWindow{
 				$b->hollow = $value;
 			});
 		}
-		if($b->mode === Brush::MODE_BRUSH && $b->getShape()->usesHeight() && $b->getType()->usesSize() && $b->getType()::ID !== BiomeType::ID){
-			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_HEIGHT), 0, $loader->config->maxSize, 1, $b->height, function(Player $player, float $value) use ($b){
-				$b->height = (int) $value;
+		if($b->mode === Brush::MODE_BRUSH && $b->getShape()->usesThreeLengths() && $b->getType()->usesSize()){
+			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_WIDTH), 0, $loader->config->maxSize, 1, $b->width, function(Player $player, float $value) use ($b){
+				$b->width = (int) $value;
+			});
+			if($b->getType()::ID !== BiomeType::ID) {
+				$this->addSlider($this->t(Translation::UI_BRUSH_MENU_HEIGHT), 0, $loader->config->maxSize, 1, $b->height, function(Player $player, float $value) use ($b){
+					$b->height = (int) $value;
+				});
+			}
+			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_LENGTH), 0, $loader->config->maxSize, 1, $b->length, function(Player $player, float $value) use ($b){
+				$b->length = (int) $value;
 			});
 		}
 		if($b->getType()->usesBlocks()){