+++ src/BlockHorizons/BlockSniper/ui/windows/PresetPropertiesWindow.php
@@ -28,12 +28,12 @@ class PresetPropertiesWindow extends CustomWindow{
 		$b->shape = $preset->properties->shape;
 		$b->type = $preset->properties->type;
 
-		if($b->mode === Brush::MODE_BRUSH && $b->getType()->usesSize()){
+		if($b->mode === Brush::MODE_BRUSH && $b->getType()->usesSize() && !$b->getShape()->usesThreeLengths()){
 			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_SIZE), 0, $loader->config->maxSize, 1, $b->size, function(Player $player, float $value) use ($b){
 				$this->preset->properties->size = (int) $value;
 			});
 		}
-		if($b->mode === Brush::MODE_BRUSH && $b->getType()->usesSize()){
+		if($b->mode === Brush::MODE_BRUSH && $b->getType()->usesSize() && !$b->getShape()->usesThreeLengths()){
 			$this->addToggle($this->t(Translation::UI_BRUSH_MENU_DECREMENT), $b->decrementing, function(Player $player, bool $value) use ($b){
 				$this->preset->properties->decrementing = $value;
 				// Set the size the brush will reset to after reaching a size of 0.
@@ -45,9 +45,17 @@ class PresetPropertiesWindow extends CustomWindow{
 				$this->preset->properties->hollow = $value;
 			});
 		}
-		if($b->mode === Brush::MODE_BRUSH && $b->getShape()->usesHeight() && $b->getType()->usesSize() && $b->getType()::ID !== BiomeType::ID){
-			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_HEIGHT), 0, $loader->config->maxSize, 1, $b->height, function(Player $player, float $value) use ($b){
-				$this->preset->properties->height = (int) $value;
+		if($b->mode === Brush::MODE_BRUSH && $b->getShape()->usesThreeLengths() && $b->getType()->usesSize()){
+			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_WIDTH), 0, $loader->config->maxSize, 1, $b->width, function(Player $player, float $value) use ($b){
+				$this->preset->properties->width = (int) $value;
+			});
+			if($b->getType()::ID !== BiomeType::ID) {
+				$this->addSlider($this->t(Translation::UI_BRUSH_MENU_HEIGHT), 0, $loader->config->maxSize, 1, $b->height, function(Player $player, float $value) use ($b){
+					$this->preset->properties->height = (int) $value;
+				});
+			}
+			$this->addSlider($this->t(Translation::UI_BRUSH_MENU_LENGTH), 0, $loader->config->maxSize, 1, $b->length, function(Player $player, float $value) use ($b){
+				$this->preset->properties->length = (int) $value;
 			});
 		}
 		if($b->getType()->usesBlocks()){